requirements.txt
# Core Dependencies
flask==3.0.0
fastapi>=0.115.6
python-multipart==0.0.6
httpx>=0.24.0

# Database
sqlalchemy==2.0.23

# AI & ML
google-generativeai==0.3.1
sentence-transformers==2.2.2
openai>=1.12.0
anthropic>=0.18.1
faiss-cpu>=1.7.4
spacy>=3.5.0
zh_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/zh_core_web_sm-3.5.0/zh_core_web_sm-3.5.0.tar.gz

# NLP & Text Processing
textblob>=0.17.1
langdetect>=1.0.9
jieba>=0.42.1
whoosh>=2.7.4

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
markdown>=3.3.4

# Config & Utils
python-dotenv==1.0.0
pydantic==2.5.2
psutil==5.9.6
requests==2.31.0
numpy>=1.21.0
scikit-learn>=0.24.2

# UI
streamlit==1.31.0
watchdog==3.0.0

# LINE Bot
line-bot-sdk>=3.7.0
aiohttp>=3.9.1

# Media Processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90
soundfile>=0.10.3

# YouTube & Google API
youtube-transcript-api>=0.6.1
google-api-python-client>=2.0.0

# Development & Testing
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
flake8>=4.0.0

pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
fastapi==0.104.1
pytest-cov==4.1.0
numpy==1.24.3
pandas==2.0.3